requirements.txt
google-auth==2.23.0
google-api-python-client==2.98.0
oauth2client==4.1.3
selenium==4.11.2
webdriver-manager==4.0.0
parsel==1.8.1
bs4==0.0.1
google_auth_oauthlib==1.1.0
timeout-decorator==0.5.0
